refactor(content_ui): remove duplicate feature graphql
