+++ frontend/plugins/content_ui/src/modules/cms/categories/graphql/queries/index.ts
@@ -1,3 +1,2 @@
 export { CMS_CATEGORIES } from './categoriesListQueries';
 export { CMS_CATEGORY_DETAIL } from './categoryDetailQuery';
-export { CMS_CUSTOM_FIELD_GROUPS } from '../../../custom-fields/graphql/queries';
