+++ frontend/plugins/content_ui/src/modules/cms/custom-types/graphql/index.ts
@@ -1,0 +1,2 @@
+export * from './mutations';
+export * from './queries';
