+++ frontend/plugins/content_ui/src/modules/cms/graphql/queries.ts
@@ -677,123 +677,3 @@ export const PAGES_REMOVE = gql`
     cmsPagesRemove(_id: $id)
   }
 `;
-
-export const CMS_CUSTOM_FIELD_GROUPS = gql`
-  query cmsCustomFieldGroupList($clientPortalId: String!) {
-    cmsCustomFieldGroupList(clientPortalId: $clientPortalId) {
-      list {
-        _id
-        label
-        code
-        clientPortalId
-        customPostTypeIds
-        customPostTypes {
-          _id
-          code
-          label
-          pluralLabel
-        }
-        enabledPageIds
-        enabledCategoryIds
-        enabledPostIds
-        fields
-      }
-    }
-  }
-`;
-
-export const CMS_CUSTOM_FIELD_GROUP_ADD = gql`
-  mutation cmsCustomFieldGroupsAdd($input: CustomFieldGroupInput!) {
-    cmsCustomFieldGroupsAdd(input: $input) {
-      _id
-      label
-      code
-      clientPortalId
-      customPostTypeIds
-      customPostTypes {
-        _id
-        code
-        label
-        pluralLabel
-      }
-      enabledPageIds
-      enabledCategoryIds
-      fields
-    }
-  }
-`;
-
-export const CMS_CUSTOM_FIELD_GROUP_EDIT = gql`
-  mutation cmsCustomFieldGroupsEdit(
-    $_id: String!
-    $input: CustomFieldGroupInput!
-  ) {
-    cmsCustomFieldGroupsEdit(_id: $_id, input: $input) {
-      _id
-      label
-      code
-      clientPortalId
-      customPostTypeIds
-      customPostTypes {
-        _id
-        code
-        label
-        pluralLabel
-      }
-      enabledPageIds
-      enabledCategoryIds
-      fields
-    }
-  }
-`;
-
-export const CMS_CUSTOM_FIELD_GROUP_REMOVE = gql`
-  mutation cmsCustomFieldGroupsRemove($_id: String!) {
-    cmsCustomFieldGroupsRemove(_id: $_id)
-  }
-`;
-
-export const CMS_CUSTOM_POST_TYPES = gql`
-  query cmsCustomPostTypes($clientPortalId: String) {
-    cmsCustomPostTypes(clientPortalId: $clientPortalId) {
-      _id
-      code
-      label
-      pluralLabel
-      description
-      createdAt
-    }
-  }
-`;
-
-export const CMS_CUSTOM_POST_TYPE_ADD = gql`
-  mutation cmsCustomPostTypesAdd($input: CustomPostTypeInput!) {
-    cmsCustomPostTypesAdd(input: $input) {
-      _id
-      code
-      label
-      pluralLabel
-      description
-      createdAt
-    }
-  }
-`;
-
-export const CMS_CUSTOM_POST_TYPE_EDIT = gql`
-  mutation cmsCustomPostTypesEdit($_id: String!, $input: CustomPostTypeInput!) {
-    cmsCustomPostTypesEdit(_id: $_id, input: $input) {
-      _id
-      code
-      label
-      pluralLabel
-      description
-      createdAt
-    }
-  }
-`;
-
-export const CMS_CUSTOM_POST_TYPE_REMOVE = gql`
-  mutation cmsCustomPostTypesRemove($_id: String!) {
-    cmsCustomPostTypesRemove(_id: $_id)
-  }
-`;
