+++ frontend/plugins/content_ui/src/modules/cms/pages/graphql/queries/index.ts
@@ -1,3 +1,2 @@
 export { PAGE_LIST } from './pagesListQueries';
 export { PAGE_DETAIL } from './pageDetailQuery';
-export { CMS_CUSTOM_FIELD_GROUPS } from '../../../custom-fields/graphql/queries';
